# Erronka2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/talde2erronka2back/Erronka2/public/api/hitzordua/{fecha}": {
            "get": {
                "tags": [
                    "Hitzorduak"
                ],
                "summary": "Hitzorduen datuak lortu",
                "operationId": "8e98b02f450efb88451f20eb33a2616c",
                "parameters": [
                    {
                        "name": "fecha",
                        "in": "path",
                        "description": "Fecha de los hitzorduak en el formato YYYY-MM-DD",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2024-02-10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Hitzorduen datuak lortu dira."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/hitzordua/ezabatu": {
            "put": {
                "tags": [
                    "Hitzorduak"
                ],
                "summary": "Hitzorduak ezabatu",
                "operationId": "751e91113ad303e7c58a1d7ebdeb49db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Ezabatu nahi den hitzorduaren IDa",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hitzordua ezabatu da."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/hitzordua/horduaHasiera": {
            "put": {
                "tags": [
                    "Hitzorduak"
                ],
                "summary": "Hitzorduen hasiera ordua editatu",
                "operationId": "f5dcb9e414f1b3bf19e47a7f6fa73fe0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "ordua"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Hitzorduaren IDa",
                                        "type": "integer"
                                    },
                                    "ordua": {
                                        "description": "Hitzorduaren hasiera ordua",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hasiera ordua editatu da."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/hitzordua/horduaAmaiera": {
            "put": {
                "tags": [
                    "Hitzorduak"
                ],
                "summary": "Hitzorduen amaiera ordua editatu",
                "operationId": "8b714f3af88acd24397247189efee982",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "ordua"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Hitzorduaren IDa",
                                        "type": "integer"
                                    },
                                    "ordua": {
                                        "description": "Hitzorduaren amaiera ordua",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Amaiera ordua editatu da."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/hitzordua/horduDisp/{data}/{hasiera_ordua}/{amaiera_ordua}": {
            "get": {
                "tags": [
                    "Hitzorduak"
                ],
                "summary": "Hitzordu libreak lortu",
                "operationId": "c10c95e94ea5b4f45cea7db7bf2adecd",
                "parameters": [
                    {
                        "name": "data",
                        "in": "path",
                        "description": "Hitzorduaren data (YYYY-MM-DD formatuan)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hasiera_ordua",
                        "in": "path",
                        "description": "Hitzorduaren hasiera ordua (HH:MM formatuan)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amaiera_ordua",
                        "in": "path",
                        "description": "Hitzorduaren amaiera ordua (HH:MM formatuan)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datuak lortu dira."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/hitzordua/txertatu": {
            "post": {
                "tags": [
                    "Hitzorduak"
                ],
                "summary": "Hitzorduak txertatu",
                "operationId": "d477a6dee71a63e67a54f1194756c7cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "izena",
                                    "telefonoa",
                                    "deskribapena",
                                    "hasiera_ordua",
                                    "amaiera_ordua",
                                    "langilea",
                                    "data"
                                ],
                                "properties": {
                                    "izena": {
                                        "type": "string"
                                    },
                                    "telefonoa": {
                                        "type": "string"
                                    },
                                    "deskribapena": {
                                        "type": "string"
                                    },
                                    "hasiera_ordua": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "amaiera_ordua": {
                                        "type": "string",
                                        "format": "time"
                                    },
                                    "langilea": {
                                        "type": "integer"
                                    },
                                    "etxekoa": {
                                        "type": "boolean"
                                    },
                                    "data": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Hitzordua txertatu da."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}